refactor: rename model/ → types/
model/ implies MVC/ORM semantics that don't apply here. types/ is the
idiomatic name in the Soroban ecosystem for contracttype-annotated
domain types (VCStatus, VerifiableCredential). Updates all import sites.
