+++ contracts/vc-vault/src/types/vc_status.rs
