+++ contracts/vc-vault/src/types/verifiable_credential.rs
